+++ oap-server/server-telemetry/telemetry-prometheus/src/main/java/org/apache/skywalking/oap/server/telemetry/prometheus/PrometheusMetricsCollector.java
@@ -30,7 +30,7 @@ import org.apache.skywalking.oap.server.telemetry.api.MetricsCollector;
 public class PrometheusMetricsCollector implements MetricsCollector {
     @Override public Iterable<MetricFamily> collect() {
         Enumeration<Collector.MetricFamilySamples> mfs = CollectorRegistry.defaultRegistry.metricFamilySamples();
-        List<MetricFamily> result = new LinkedList<>();
+        List<MetricFamily> result = new ArrayList<>();
         while (mfs.hasMoreElements()) {
             Collector.MetricFamilySamples metricFamilySamples = mfs.nextElement();
             List<MetricFamily.Sample> samples = new ArrayList<>(metricFamilySamples.samples.size());